backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
pydantic>=2.5.0
pydantic-settings>=2.4.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiohttp==3.9.1
google-generativeai==0.3.2
protobuf>=5.26.1,<6.0
